教务查询问题解决了
问题解决

+++ src/main/java/com/newer/sc/common/entity/Student.java
@@ -1,5 +1,7 @@
 package com.newer.sc.common.entity;
 
+import java.math.BigInteger;
+
 import java.sql.Date;
 
 import org.springframework.stereotype.Component;
@@ -12,7 +14,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class Student {
 	
-	int sid;
+	BigInteger sid;
 	
 	String sname;
 	
@@ -32,11 +34,11 @@ public class Student {
 	
 	Major major;
 
-	public int getSid() {
+	public BigInteger getSid() {
 		return sid;
 	}
 
-	public void setSid(int sid) {
+	public void setSid(BigInteger sid) {
 		this.sid = sid;
 	}
